增加如家酒店评论抓取

+++ rarone-core/src/main/java/com/codeclen/rarone/core/instance/bestwehotel/extract/BestwehotelCommentExtractor.java
@@ -46,7 +46,7 @@ public class BestwehotelCommentExtractor extends AbstractRoomExtractor {
                                     comment.setReplyTime(replyTime);
                                 }
                             }
-                            comment.setEntireScore(commentObj.getInteger("entireScore"));
+                            comment.setEntireScore(commentObj.getDouble("entireScore"));
                             res.add(comment);
                         }
                     }
